+++ src/IntervalSet.php
@@ -109,7 +109,7 @@ final class IntervalSet implements TemporalSet
      *
      * @return list<non-empty-string>
      */
-    public function allFormatted(IntervalFormat $format = IntervalFormat::Iso8601StartDuration, ?Unit $unit = null): array
+    public function formatAll(IntervalFormat $format = IntervalFormat::Iso8601StartDuration, ?Unit $unit = null): array
     {
         return array_map(static fn (Interval $item): string => $item->format($format, $unit), $this->items);
     }