+++ src/TaskSet.php
@@ -89,7 +89,7 @@ final class TaskSet implements TemporalSet
      *
      * @return list<non-empty-string>
      */
-    public function allFormatted(IntervalFormat $format = IntervalFormat::Iso8601StartDuration, ?Unit $unit = null): array
+    public function formatAll(IntervalFormat $format = IntervalFormat::Iso8601StartDuration, ?Unit $unit = null): array
     {
         return array_map(static fn (Task $item): string => $item->format($format, $unit), $this->items);
     }